Split Ollama into a smaller base version for the browser, and a subclass that supports Node
